backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blis==1.0.1
breadability==0.1.20
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0.tar.gz#sha256=e8813936311a536d2ce7df155817674f20e2a3896dd0e875367446049a00e0dd
exceptiongroup==1.2.2
fastapi==0.95.0
filelock==3.16.1
fsspec==2024.9.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.9
Jinja2==3.1.4
joblib==1.4.2
keybert==0.8.5
langcodes==3.4.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.1
nltk==3.9.1
numpy==2.0.2
packaging==24.1
pandas==2.2.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.0.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.28.1
psycopg2==2.9.9
psycopg2-binary==2.9.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==1.10.4
pydantic_core==2.23.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2024.2
PyYAML==6.0.2
rake-nltk==1.0.6
regex==2024.9.11
requests==2.28.1
rich==13.9.2
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.26.1
sympy==1.13.3
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.24.0
typer==0.4.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
wasabi==0.10.1
weasel==0.4.1
wrapt==1.16.0
